Refactor: Centralize probing logic and improve model wrapper robustness
